convert all submodules into class structure
